Improve code as per code review
