+++ src/module.indexer/_module.ts
@@ -10,7 +10,7 @@ import { TransactionVinIndexer } from '@src/module.indexer/model/transaction.vin
 import { TransactionVoutIndexer } from '@src/module.indexer/model/transaction.vout'
 import { VoutFinder } from '@src/module.indexer/model/_vout_finder'
 import { IndexStatusMapper } from '@src/module.indexer/status'
-import { OracleStatusIndexer } from '@src/module.indexer/model/oracle.status'
+import { OracleAppointedIndexer } from '@src/module.indexer/model/oracle.appointed'
 
 @Module({
   providers: [
@@ -25,7 +25,7 @@ import { OracleStatusIndexer } from '@src/module.indexer/model/oracle.status'
     TransactionIndexer,
     TransactionVinIndexer,
     TransactionVoutIndexer,
-    OracleStatusIndexer
+    OracleAppointedIndexer
   ]
 })
 export class IndexerModule {
