+++ src/module.indexer/model/_main.ts
@@ -11,7 +11,7 @@ import { JsonRpcClient } from '@defichain/jellyfish-api-jsonrpc'
 import { RawBlockMapper } from '@src/module.model/raw.block'
 import { NotFoundIndexerError } from '@src/module.indexer/error'
 import { blockchain as defid } from '@defichain/jellyfish-api-core'
-import { OracleStatusIndexer } from '@src/module.indexer/model/oracle.status'
+import { OracleAppointedIndexer } from '@src/module.indexer/model/oracle.appointed'
 
 /**
  * This is a deterministic log based indexer.
@@ -30,7 +30,7 @@ export class MainIndexer {
     private readonly transaction: TransactionIndexer,
     private readonly transactionVin: TransactionVinIndexer,
     private readonly transactionVout: TransactionVoutIndexer,
-    private readonly oracleStatus: OracleStatusIndexer
+    private readonly oracleAppointed: OracleAppointedIndexer
   ) {
     this.indexers = [
       block,
@@ -40,7 +40,7 @@ export class MainIndexer {
       transaction,
       transactionVin,
       transactionVout,
-      oracleStatus
+      oracleAppointed
     ]
   }
